Changes to adapt to the global handling exceptions

+++ Todo.Application/Services/TodoItemService.cs
@@ -5,6 +5,7 @@ using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using Todo.Domain.DTOs;
 using Todo.Domain.Entities;
+using Todo.Domain.Exceptions;
 using Todo.Domain.Interfaces.Repositories;
 using Todo.Domain.Interfaces.Services;
 
@@ -28,13 +29,16 @@ namespace Todo.Application.Services
         public async ValueTask<TodoItem> CreateAsync(TodoItem item)
             => await _todoItemRepository.CreateAsync(item);
 
-        public async ValueTask<TodoItem> UpdateAsync(long id, TodoDTO itemDTO)
+        public async ValueTask<TodoItem> UpdateAsync(long idTodo, TodoDTO itemDTO, long idUser)    
         {
-            var obj = await _todoItemRepository.GetByIdAsync(id);
+            var obj = await _todoItemRepository.GetByIdAsync(idTodo);
 
             if(obj == null)
-                throw new Exception("NotFound");
+                throw new NotFoundException("Not Found");
             
+            if(CheckTodoBelongsUser(obj, idUser))
+                throw new UnauthorizedException("Not authorized");
+
             obj.Title = itemDTO.Title;
             obj.Description = itemDTO.Description;
 
@@ -43,14 +47,22 @@ namespace Todo.Application.Services
             return obj;
         } 
 
-        public async ValueTask<TodoItem> DeleteAsync(long id)
+        public async ValueTask<TodoItem> DeleteAsync(long idTodo, long idUser)
         {
-            var todo = await _todoItemRepository.GetByIdAsync(id);
+            var todo = await _todoItemRepository.GetByIdAsync(idTodo);
 
             if(todo == null)
-                throw new Exception("NotFound");
+                throw new NotFoundException("NotFound");
             
+            if(CheckTodoBelongsUser(todo, idUser))
+                throw new UnauthorizedException("Not authorized");
+
+
             return await _todoItemRepository.DeleteAsync(todo);
         }
+
+
+        private bool CheckTodoBelongsUser(TodoItem todo, long userId)
+            => todo.IdUser == userId;
     }
 }
